#161 Make background event errors go away
