+++ extension/ts/main/extensions/util/onLogged.ts
@@ -4,7 +4,8 @@ import {createIconButton} from "../../ui/button";
 import {isAuthorized} from "../author/util/isAuthorized";
 import {showToast, ToastType} from "../../ui/toast";
 import {loaderOverlaid} from "../../ui/loadingIndicator";
-import {BackgroundEvent, onBackgroundEvent} from "../../../common/backgroundEvent";
+import {BackgroundEvent} from "../../../common/backgroundEvent";
+import {onBackgroundEvent} from "../../util/onBackgroundEvent";
 
 const BAD_BROWSER_INFO_URL =
 	"https://github.com/braxtonhall/library-thing/blob/main/docs/librarian/authors.md#prerequisites";
